requirements.txt
# This file is automatically generated by pip-compile
 
 # To update it, run: pip-compile requirements.in
# To install the requirements, run: pip install -r requirements.txt
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
bettercam==1.0.0
black==25.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.4
configparser==7.2.0
contourpy==1.3.0
coverage==7.9.1
cryptography==42.0.8
cycler==0.12.1
Cython==3.0.10
dill==0.3.8
distro==1.9.0
easygui==0.98.3
easyocr==1.7.1
exceptiongroup==1.3.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
glfw==2.7.0
GPUtil==1.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==2.10
imageio==2.35.1
iniconfig==2.1.0
inputs==0.5
Jinja2==3.1.4
jiter==0.10.0
keyboard==0.13.5
kiwisolver==1.4.7
lazy_loader==0.4
logger==1.4
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
mss==6.1.0
mypy_extensions==1.1.0
networkx==3.3
ninja==1.11.1.1
Nuitka==2.4.8
numpy==1.26.4
nvidia-cuda-runtime-cu12==12.5.82
onnx==1.16.1
onnxconverter-common==1.14.0
onnxoptimizer==0.3.13
onnxruntime-directml==1.19.0
onnxruntime-gpu==1.19.0
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
ordered-set==4.1.0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
Pillow==9.5.0
platformdirs==4.3.8
pluggy==1.6.0
protobuf==3.20.2
psutil==6.0.0
py-cpuinfo==9.0.0
pyAesCrypt==6.1.1
pyclipper==1.3.0.post5
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.5.2
pyparsing==3.1.4
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyreadline3==3.4.1
pyserial==3.5
pyserial-asyncio==0.6
pytest==8.3.5
pytest-cov==6.2.1
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.2
pywin32==302
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
sympy==1.12.1
tensorrt==10.11.0.33
tensorrt_cu12==10.11.0.33
tensorrt_cu12_bindings==10.11.0.33
tensorrt_cu12_libs==10.11.0.33
termcolor==3.1.0
tifffile==2024.9.20
tomli==2.2.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.66.5
tradingeconomics==4.4.9
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.2.62
ultralytics-thop==2.0.8
# Editable Git install with no remote (UNKNOWN==0.0.0)
-e g:\dorianspcstuff\aim_assist_app
urllib3==2.2.2
vgamepad==0.0.8
websocket-client==1.8.0
zstandard==0.23.0
